valgrind: Updated valgrind recipe
Change 1:
Added sysroot-librt to DEPENDS/RDEPENDS due to it being found as not found
in config.log

Change 2:
Added sysroot-dbg to RDEPENDS, due to the debug information being needed
in the valgrind tool: memcheck (the default used tool). The
ld-2.11.1.so created by crosstool-ng is stripped and the debug information
is needed for valgrind

change 3:
Added a patch that removed a segmentation fault when run on ppc32.
problem:
"The root problem is that the newer gcc was not generating
the appropriate code to long jump back to the correct location.
Instead, the long jump target was indeterminate, which in this case,
resulted in a segmentation fault."
see: http://bugs.kde.org/show_bug.cgi?id=259977