ntp: Fix not installing ntpd init script, remove local system clock fallback
The comment in the ntp.conf file "Using local hardware clock" is not correct
it will use the local system clock, i.e. the clock ntpd is trying to keep
correct. Read comments in ntp-4.2.6p3/ntpd/refclock_local.c for reasons to
use this clock. Removing the line has the added benefit that filesystems
using ntpdate can be dropped and call to ntpdate <servers> can be replaced
with "ntpd -q -g".