binutils: Fixup -dbg packaging and add fixup some RDEPENDS
Esben: The RDEPENDS stuff in this commit is not really the proper solution,
but is needed for now.  We properly should work on getting binutils
binaries to be built as part crosstool-ng, so that we can use the
right binutils libraries, while also avoiding building all this stuff
twice for no good reason.