base-initramfs: Refactor to new initramfs-image and drop .inc file
Use the new initramfs-image class which does not require inherit of
rootfs.

The .inc file does nothing but inherit, so we don't want to keep it as
a part of the OE-lite/base API.

Actually, I think we should deprecate any use of include of .inc files
across recipe directories and especially layers.