overlay-files: Add recipe and class for managing file overlays
This can fx. be used for keeping modification of configuration files in
the rootfs in a different filesystem, which will then be retained when
replacing the rootfs.  This is a very simple solution, and any changes
that needs to be done to such overlayed files after upgrading the rootfs
is not handled by this solution.

(closes #367)