e2fsprogs: Add defines for mapping old non-API uuid functions
The use of uuid__generate_random() and uuid__generate_time(), which
were never part of the uuid official API, is now broken, as they are
renamed to __uuid_generate_random() and __uuid_generate_time() in
newer util-linux versions.

This was previously added in util-linux by a patch, which has been
dropped.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>