util-linux: Fixup install dir handling
With this, the binaries will end up where we want them.  Some of them
has hardcoded the old paths still, so some amount of patching should
also be added for them to actually work.

But at least, we can now build and package it, and the libraries should
be fine (which is what we need for sdk builds).

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>

(cherry picked from commit b32b9ae658c8077cb17969cc0fe5232e9883c31b)