directfb: Fix rebuilds and re-enable parallel build
The problems observed previously is actually caused by the distclean make
target, which is removing a bunch of .c source files which are not being
built when using the tarball release.  By using VPATH the problem is
worked around, as the files are left intact in srcdir.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>

(cherry picked from commit 46649a3bc77c24f69a8ef60767944ce121461565)