binutils: Raise priority to fix multiple provider problem
As both gcc recipe in OE-lite/core and this recipe provides libbfd, we
need to ensure that one of them is preferred over the other.  For now,
we choose this binutils recipe.

When a proper layer priority mechanism is implemented, this will typically
become default also, as base will likely be set at a higher priority
than core.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>