util-linux: Added version 2.25.2
Previously, in 2.24, --without-python was default in configure and had to
be manually enabled. In 2.25.2, it is by default --with-python and has to
be manually disabled.

New utils added: blkdiscard, nologin, swapoff, last, lastb, lslogins,
mesg, nsenter, prlimit, uname26. Some was already present in 2.24, but
they are all splitted out now.

Added new library: smartcols.

Due to the addition of utilities and the library, a lot of dependency
requirements has been added in 2.25.2.

Signed-off-by: Christian Sørensen <christian.braunersorensen@prevas.dk>