busybox: Add s6-rc integration of hwclock
This adds an s6-rc service called rtc, which reads hardware clock (RTC)
on start, and writes it on stop.  It can be handy as startup and shutdown
service.

If /etc/hwclock/UTC file exists and does contain something else than the
string "1", the time hwclock is assumed to contain local time. Otherwise,
it is assumed to contain UTC time.

If /etc/hwclock/DEVICE file exists, it specifies the RTC device to use.
Otherwise, /dev/rtc is used.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>