s6-init: Add s6-rc bootstrap procedure
When booting without a compiled service database, assume that the
first compiled database is not complete, and use it for bringing up the
bootstrap service (bundle), and then recompile a new service database,
and first then bringup the default service (bundle).

This allows for using s6-init hook functions that relies on other services
to be up.

To activate this, include a service called 'bootstrap' in your /etc/rc.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>