s6-init: Sanitize syslog file format
This changes the format used in /var/log/syslogd to match the format used
in all the other /var/log/* dirs.

The TIMESTAMP field coming from syslog is discarded.  It is presented in
locale timezone, which is pretty useless in a global world, and does not
contain information which year the date belongs to!  All lines are prefix
with tai64 timestamps by s6-log instead.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>