libunwind: make it a little neater
Utilizing the new AUTO_PACKAGE_LIBS_LIBPREFIX, along with the existing
PROVIDEPREFIX and PKGPREFIX, we can arrange that the recipe generates
packages "libunwind-setjmp", "libunwind-coredump", etc. which provide
the .so files of the same name.

While at it, make sure this is only compatible with arm for now. The
proper fix would be to extract the proper arch string and use that
instead of the hardcoded "arm".

Another nice-to-have is a use-flag for enabling lzma.

For now, let's just be content with the fact the packageqa passes,
whatever that means.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>