busybox: stty: Allow baudrates up to 4000000
Probably due to memory usage the baudrate is coded to make it possible to
store it in a short.
This is done by dividing high baudrates by 256 (shift 8 bits) and raising
MSB to signal the division.
This way baudrates being dividable with 256 and not greater than 8388352
can be stored.

To make support for the 'odd' baudrates of 1500000, 2500000 etc bit 14
(MSB -1) has bin chosen to signal a divisor of 250.

Drawback is that coded max, after being divided by 256, must now not
exceed 0x3FFF causing maximum baudrate to be 4194048
or 4095750 for baudrates dividable by 250