barebox: remove recipe
The current barebox recipe is a little flaky (the .inc file points to
a version-specific SRC_URI, and the .oe file overrides that with a
SRC_URI pointing to a commit which is not v2012.02, but actually just
two commits shy of v2012.03). It also causes problems for at least one
downstream user of oe-lite (preventing "oe bake world" from even
parsing) because of the default dependency on the existence of a
barebox-defconfig recipe. Nuke it.