libxml2: actually (try to) use oe-lite python
There are two reasons the python_setup function does nothing: It is
never run because of the prefuncS typo. And even if it did get run, it
would be in a different shell instance than where the do_configure
function runs.

Try to fix this by just defining the variables and setting proper
flags ensuring they're only emitted to the do_configure task.