lzma: remove recipe
This recipe only builds a /usr/bin/lzma binary and a static library
/usr/lib/liblzma.a. The latter is rather useless without header files,
and the former might as well be provided by the xz recipe.

One problem with keeping this 7-zip based lzma recipe around is that
its version (9.20) is higher than the xz-utils recipe (5.2.2), which
means that there's no (usable) liblzma we can depend on.