fuse: Fix build when running with mknod privileges
Without this fix, you will end up creating ${D}/dev/fuse when building
with mknod privileges. And as /dev/fuse is not included in any package,
do_split barfs.

Yes, you probably shouldn't do this.  But if (when) done, there is no
good reason to fail in this odd way.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>