sqlite3: move code to .inc file
The existing SRC_URI and S settings in the .inc file are useless and
confusing, since they get unconditionally overridden by the only recipe
file including it. But we might as well move the PV_MANGLED logic from
the .oe file to the .inc file. While at it, also move the LICENSE - the
sqlite website is pretty clear that the code is and always will be
public domain.