busybox-configure.inc: replace "0"/"1" USE flag values by False/True
Many of the USE_busybox_* flags really are boolean flags, but some
actually use the string or integer value of the flag. For example,
USE_busybox_syslogd_rotate_n is supposed to be an integer, and just
happens to default to 1.

Since the OE-lite lexer currently treats the token False exactly the
same as the string "0", and similarly for True/"1", we can make the
recipe a little more self-documenting by using False/True to initialize
the boolean flags, while leaving strings for the rest.