mobile-broadband-provider-info: new recipe
This must set a number of records. As if ${PN} wasn't long enough by
itself, the releases use the YYYYMMDD format. Furthermore, we need to
use autotools-autoreconf, and all this just to get a single .xml file
and an associated .dtd into /usr/share.

This is a prerequisite for ofono.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>