dosfstools: inherit pkgconfig
Without this, the configure script ends up using the host pkgconfig,
which at least for me ends up finding libudev, which then causes the
compile to fail since we don't have that.

We should probably also add some explicit USE_dosfstools_udev flag to
control this explicitly, but for now it suffices to use the proper
pkg-config.