boost.inc: remove stray do_fetch prefunc
First of all, the addhook statement has the side-effect of setting the
emit flag to "" for the function in question, so add_deps is not
available when do_fetch runs, so this does nothing. (When asked to run a
non-existing function, OE-lite substitutes an always-successful noop
function.) I also fail to see what setting an RDEPENDS_ variable has to
do with do_fetch.