opencv: fix mingw build
libdl is only listed as a package dependency for the package opencv, not
as a build dependency. Doing "oe bake opencv" works fine, but trying to
build anything that DEPENDS on opencv gives

FATAL: No provider for machine:libdl (needed by machine:opencv)

in the mingw build. Move the libdl dependency to the DEPENDS_LIBC
variable which is cleared for the mingw build. While maybe not the
correct/best fix, both native:opencv and machine:opencv pass packageqa
for both mingw and arm-926ejs.