util-linux: ncurses fixups
When trying to compile with USE_ncurses set, I hit lots of weird
errors. I think part of the problem is that --with-ncursesw is 'auto' by
default, so it is not enough to disable that in the !USE_ncurses case,
it very much also has to be explicitly disabled when we want to use the
--with-ncurses option. Otherwise, the configure script tries
to run ncursesw-config, which it happened to pick up from the host, and
then it obviously fails when it can't link with libncursesw.

But we also have to explicitly disable widechar support to convince the
configure script that --with-ncurses is ok, otherwise one gets a
non-helpful

  wide-char support enabled, but non-wide ncurses selects

For me (with MACHINE_CPU = "arm-926ejs"), the end result passes
packageqa with and without USE_ncurses set.