dosfstools: add compat symlinks
We're not currently installing e.g. mkfs.vfat despite the provide of
util/mkfs.vfat. But dosfstools can easily be convinced to create these
legacy symlinks.

It's easy enough to get those symlinks packaged with the binary they
point to. Unfortunately, the dots are translated to dashes by
auto-package-utils, so we have to provide both a name for the util/foo
part as well as the name of the symlink itself, and at the end of the
day, we also have to manually add provides which do include the
dot (since some recipes may be explicitly depending on util/mkfs.vfat).