ntpd: change stopdaemon logic
The current stopdaemon only stops one specific ntpd process (the one
listed in the pid-file). However, ntpd may start helper child
processes (apperently especially if there's no network available when it
starts). If we do not kill those as well, a subsequent startdaemon will
fail since busybox' start-stop-daemon detects that there's already a
process running the /usr/bin/ntpd executable. While the helper children
usually themselves detect that their parent is gone after a while, this
doesn't help the case of the restart or reload commands.

So change the logic in stopdaemon to kill all processes running the ntpd
executable. While at it, put the path in a variable to help avoid typos.