coreutils: clean up dependencies
It is reasonable to expect that if one RDEPENDS on "coreutils", one
gets all the utils provided by the coreutils recipe. This is also true
when one doesn't set the USE_coreutils_singlebin flag. However, with
that flag, the dependencies are turned inside-out, and all the
auto-packages now rdepend on the main package, which is where the
coreutils binary is packaged. Hence in that case, one gets _only_ that
binary, but none of the "aliases" for that binary.

Instead, when we create a single binary, put that in its own
non-auto-utils package, and make all the auto-packages rdepend on that
one. Let the main package coreutils unconditionally be an empty
package which always just rdepends on all the auto-packages.

While at it, set the appropriate DEPENDS/RDEPENDS for the various
packages - after getting rid of libiconv, there are very few that need
anything beyond libc.

This now passes packageqa with and without coreutils_singlebin set.