wtmptrim: new recipe
This is a small utility for trimming /var/log/{btmp,wtmp} files to
prevent them from filling up the disk over time (each login/failed login
adds about 384 bytes (depending on architecture)). It tries to
ensure that logins done while it is running are also preserved, but
doesn't quite guarantee it.

It's also somewhat paranoid in that, by default, if something goes wrong
after we've opened and locked the wtmp file, we truncate it to zero size
- this is to handle situations where we couldn't create the temporary
copy due to ENOSPC (which is what this is all about preventing), or
where the file is corrupted anyway (doesn't have a size which is a
multiple of the record size).