s6-init: create btmp with 0600 permissions
Currently, openssh refuses to write to the btmp file (which is otherwise
supposed to have a record of failed login attempts) with

  Excess permission or bad ownership on file /var/log/btmp

The rationale is that a common error is to type the password as the
username, and the username is part of the btmp record, so openssh
refuses to write this information to a file readable by anyone but
root. Hence make sure that it's created with 0600 permissions.

While at it, make the three clauses more symmetric.