vsftpd: add s6 init support
vsftpd has no way of getting it to log to stdout/stderr. Instead, we
have to tell it to write to the fifo s6-log is reading from. Setting
these on the command line means one doesn't have to deal with this in
the .conf file. It doesn't actually appear to ever write anything to
stdout/stderr, so one might want to redirect those to /dev/null, but I
suppose having them point at the fifo doesn't hurt.

Note that vsftpd reads the default configuration file (/etc/vsftpd.conf)
after parsing the command line if (and only if) no configuration files
were given on the command line.

The one logging-related thing one might do in the .conf file is to set
xferlog_std_format=YES if one wants wu-ftpd compatible logs (whatever
that means).