openvpn: add s6 init support
Use the rc.hooks mechanism to create one openvpn service for each
configuration file found in /etc/openvpn/. For now, let them all use the
same logger; that should be easy enough to change if required.

I've tested that the services are created as expected for each .conf
file, but I don't have an example .conf to test whether the service
actually works (I do get complaints in the log about an incomplete
configuration, though, so the logging seems to work).