dhcp: add knob for patching out syslog calls
This adds a flag (USE_dhcp_disable_syslog) which one can use to include
a patch that #if 0s out all syslog-related calls. The patch also removes
all includes of syslog.h, so that if it is forward-ported to a new
version which introduces other syslog-calls, the build hopefully fails
due to missing function declarations.

One exception is stables.c, where syslog.h is only used to build a table
of known facilities, which in turn is only used for parsing the
log-facility configuration option. While that option has no effect when
we're not actually calling syslog, it seems a little harsh to have an
existing configuration file fail to parse.

Note that dhcp links with libisc from the bind package, and some
functions in that library does syslog calls, so this
doesn't (necessarily) completely prevent dhcpd from spamming syslog - at
least, objdump on dhcpd still shows syslog calls, but whether they are
all in dead code I don't know.