overlay-files: allow symlinking to a directory
If /etc/openvpn already exists as a symlink to
/etc/openvpn.default, doing

  ln -sf /some/where /etc/openvpn

ends up creating the link where inside /etc/openvpn.default, rather than
making /etc/openvpn itself a symlink to /some/where. Add -n to avoid
dereferencing the target.

However, if /etc/openvpn happens to be a plain directory, this would
still end up creating the symlink inside that. So also add -T to force
ln to treat the linkname as-is.

Note that (a) the busybox help text for -T is actively wrong ("-T 2nd
arg must be a DIR"), (b) with busybox (but not coreutils), -n is also
necessary to handle the case of linkname being a symlink to an existing
directory.