ntp: fix default configuration
When no network is available when ntpd starts (more specifically, when
DNS queries fail), ntpd spawns a child to do future queries. When ntpd
detects a change in interfaces, it wakes up the child to make it attempt
resolving again. However, the default ntp.conf interferes with the
communication, one gets output like

 8 Aug 11:33:08 ntpd[347]: Deferring DNS for 0.pool.ntp.org 1
 8 Aug 11:33:08 ntpd[347]: Deferring DNS for 1.pool.ntp.org 1
 8 Aug 11:33:08 ntpd[347]: Deferring DNS for 2.pool.ntp.org 1
 8 Aug 11:33:08 ntpd[347]: Deferring DNS for 3.pool.ntp.org 1
 8 Aug 11:33:08 ntpd[375]: signal_no_reset: signal 17 had flags 4000000
 8 Aug 11:33:11 ntpd[347]: Listen normally on 6 wlan0 fe80::6f0:21ff:fe2f:8dc3 UDP 123
 8 Aug 11:33:11 ntpd[347]: peers refreshed
 8 Aug 11:33:11 ntpd[347]: new interface(s) found: waking up resolver
 8 Aug 11:35:08 ntpd[347]: Listen normally on 7 ppp0 10.0.0.193 UDP 123
 8 Aug 11:35:08 ntpd[347]: peers refreshed
 8 Aug 11:35:08 ntpd[347]: new interface(s) found: waking up resolver
 8 Aug 11:35:10 ntpd[375]: DNS 0.pool.ntp.org -> 92.246.24.228
 8 Aug 11:35:10 ntpd[347]: process_private: failed auth mod_okay 0
 8 Aug 11:35:10 ntpd[375]: ntp_intres.request: permission denied

I'm not entirely sure what this means, but from reading
http://doc.ntp.org/4.2.6p5/accopt.html, I think what is happening is
that the child is trying to modify the parents configuration to include
the server 92.246.24.228, but that's denied by the nomodify (or maybe
nopeer) default restriction.

This misconfiguration prevents the code in ntpd which is supposed to
handle interfaces from coming and going from working. Fix that by adding
a line that doesn't restrict localhost from anything. At the same time,
we might as well tighten the default restrictions somewhat more - if
some BSP actually wants to answer ntp queries or allow ntpd to be
configured remotely, they probably know what they're doing and can
supply their own ntp.conf.