util-linux: teach hwclock -l as alias for --localtime
It's a little easier to write scripts that will work with both busybox'
and util-linux' hwclock if they understand the same set of
options. Currently, util-linux only understands --localtime, but long
option support may be compiled out of busybox, in which case that only
understands -l. This just ensures that util-linux also understands -l,
which also makes it symmetric with the -u/--utc option.