util-linux: avoid consistent .5 second loss with hwclock --systohc
This adds a patch to util-linux to override a hard-coded assumption
about the hardware clock's behaviour when being set. On the board I'm
currently working on, passing --delay 0.0 to hwclock --systohc now
makes the board come up with a system clock that is consistently less
than 0.1s from NTP time, and usually better than 0.02s.