openvpn: create bundle consisting of all openvpn@* services
In addition to creating the openvpn@X services for each X.conf, also
create a bundle simply called openvpn that contains all of these
services.

The lack of -p in the mkdir command is deliberate; this ensures that
we're actually creating a new bundle and do not clobber some other
hook's output (the caller of the hook script is supposed to have created
the ${1} directory). We still need to create an empty contents file
explicitly in case there are no .conf files.