util-linux: allow compile-time changing of default --delay option
While the patch added in f485a62 made it possible to avoid losing
precision when doing hwclock --systohc, in practice it turns out to be
cumbersome to ensure that every call in an entire BSP passes the
appropriate --delay option.

So allow setting USE_hwclock_delay in e.g. a machine configuration. This
could probably replace the patch added in f485a62 (i.e., simply change
the hard-coded default, without adding a command-line switch to override
it), but partly because that's already merged, partly because there
might be hardware out there with two different RTCs that need different
--delay args, I'm just adding this on top.