s6rc-sanity.oeclass: s6 services sanity checking
Booting an image where the s6 service tree has a dependency cycle (so
that s6-rc-compile fails) is rather annoying - in at least one case, the
board was hanging and it took me quite a while to figure out what was
wrong and to get the board back into a usable state.

This is a simple attempt at doing a little build time sanity
checking. The rc.hooks concept makes this a little complicated; running
s6-rc-compile on /etc/rc/ alone is almost guaranteed not to work due to
missing service names, so we have to fake the dynamic bundles. For now,
simply assume that each hook script generates a bundle by the same
name. This shouldn't hurt (unless one of those happen to collide with a
name in /etc/rc).

As an example, this is what I'd gotten instead of a broken image on
target if this had been in place:

s6-rc-compile: fatal: cyclic service dependency reached from ntpd and involving ntpd
...
CRITICAL: bake failed: error: 1