dhcp: prevent fifo creation race
dhcpd depends on dhcpd-log (and net). As part of dhcpd-log's run script,
the fifo /run/dhcpd-log.fifo is created. However, dhcpd's run script
contains

redirfd -w 1 /run/dhcpd-log.fifo

Since dhcpd-log is considered started by s6 as soon as the fork+exec is
succesful, it may then proceed to starting dhcpd. Here's a race: if
dhcpd's run script reaches the redirfd line before the fifo is actually
created, we'll get a regular file, which results in three nasty
behaviours: When the s6-log finally starts, it's stdin is connected to
that regular file, so it reaches EOF rather soon, causing it to exit,
only to get respawned by its supervisor, ad infinitum. But even in its
short life time, the s6-log process has copied the contents from the
"pipe" to the log file, so the log files get filled with copies of the
same lines over and over. Meanwhile, dhcpd happily chugs along,
occasionally writing a new line to its log file, which slowly fills the
tmpfs at /run. So we have

(1) an infinite stream of fork+execs, keeping a CPU rather busy

(2) filling the log files (presumably in permanent storage) with junk,
wearing down the flash

(3) eventually exhausting ram by the regular file in /run.

Prevent this by having the dhcpd-log service use a notification-fd, so
that dhcpd is not spawned until we know the fifo is created. While at
it, make it all a bit more robust.

(a) Don't just check whether /run/dhcpd-log.fifo exists, but insist that
it is a named pipe. If the test fails due to the path existing but being
some other kind of file, the mkfifo will fail. We could throw in an 'rm
-f', but there's a bug in the system somewhere if something else has
claimed the name /run/dhcpd-log.fifo.

(b) Use -c instead of -w to prevent accidentally creating a regular
file. It is a bug if, despite the notification-fd, the dhcpd.run script
is called before the fifo exists, so we shouldn't proceed silently.