dhcp: fix openssl dependency
This is not pretty, but the best fix I could think of.

With USE_bind_openssl (which is set by default), bind and dhcp get
compiled with ssl bindings. However, it then turns out that dhcpd fails
to start with a useless "Can't initialize context: failure"
message. strace'ing shows that it tries to open()
/usr/lib/engines/libgost.so, which is packaged in openssl-engines. Since
this ultimately comes from some dlopen() call deep in openssl, no
packageqa could detect this at build time.

So instead of just depending on libssl, depend on the whole openssl
package (which itself happens to be empty but has build and runtime
dependencies on its subpackages).

This does make it harder to swap in libressl as libssl, but to fix that
I think we'd need some way of telling that "libssl is provided by
libressl" or "libssl is provided by openssl". Also, the use flag is called
bind_openssl, not just bind_ssl, so it's reasonable to have that mean a
dependency on openssl.