dhcp: fix and extend run script
It turns out that dhcpd is unable to create the leases file by itself,
so we have to ensure that at least an empty file exists.

While at it, allow the user to configure the location of the config and
lease files and which interfaces to listen on via an envdir.

Note that the lease file is supposed to persist across a reboot, to
prevent the same ip address being handed out to different machines. If
/var is tmpfs, one should thus point LEASE_FILE somewhere else.