netbase: avoid ifplugd writing to syslog
We're already setting up stdout/stderr to go to the network-log service,
there's no point in having ifplugd write the same info to syslog. So
just as it makes sense to hardcode the -n (don't "daemonize") in the s6
run script, also hardcode -s (don't syslog), regardless of what $ARGS
happen to contain.