skalibs: actually honour USE_skalibs_ipv6
The EXTRA_OECONF variable was never passed to configure (or anything
else), so USE_skalibs_ipv6 had no effect. We're also going to need
to append to EXTRA_OECONF in specific .oe files, so add it where one
expects it to be used.

While at it, always append to RECIPE_FLAGS, just in case some version
.oe file has already declared a flag.