skalibs: add version 2.6.0.1
The sysdeps and .lib files are taken from
<https://github.com/CMB/skaware-void-sysdeps>. Some comments on those:

The musl and glibc variants are almost identical. They differ in

- whether they claim support for revoke(2) (glibc=yes, but in practice
  only an ENOSYS stub is provided, so "no" is a more {truth,use}ful
  value);

- whether timer_create and friends require -lrt (musl=no, but it does no
  harm);

- the reported value of sizeof(ino_t) for arm (musl: 8, glibc: 4). That
  one is slightly interesting and worrisome. Nothing in skalibs or other
  skaware software seems to refer to the macro SKALIBS_SIZEOFINO [why
  would it when sizeof(ino_t) can just as well be used?], but there are
  some skalibs interfaces that take an "ino_t *" argument, so obviously
  caller and callee need to agree on what type ino_t really is. With
  glibc, on a 32 bit platform, that depends on whether
  -D_FILE_OFFSET_BITS=64 is used. It's not unique to skalibs, but if
  progA depends on libB, and ino_t's (or off_t's) are exchanged between
  them, they need to be compiled with the same setting of
  -D_FILE_OFFSET_BITS. We really should have a global USE flag to set
  that in c.oeclass. For now, since most recipes do not set
  _FILE_OFFSET_BITS, go with the legacy value of 4.

Also, all the original sysdeps files claim support for
getrandom(2). It's reasonable to assume that nobody wants to build new
BSPs with anything as old as 3.16 (which is more than three years old by
now), and our kernel-headers recipe certainly defines SYS_getrandom, but
I had/have to build the native recipe on a build host which does not
have such new headers. Hence I set getrandom=no on x86.