execline: add version 2.3.0.3
The old version's configure script did accept --build (but ignored it,
relying on the env var CROSS_COMPILE to find a cross compiler); in the
new version, the configure script compates $build to $host and set
cross=${host}- if they differ. To make that heuristic work for native,
we need to pass a value for --build as well as for --host.

The old version of execline won't compile against the new libskarnet
version, so a distro would have to carefully set various "prefer ..." if
the old version is to be used.