skaware: add -g to CFLAGS
Having debugging symbols (-g) available makes it much easier to get
valgrind to tell us where the problems are, and we're stripping the
binaries anyway, so it's not causing any extra memory pressure. Of
course, that stripping means that for this to be of any use, one must
also add the -dbg packages to the rootfs.