skalibs: allow building from git source
This makes it easier to try out a fix from upstream before it becomes
part of an actual release, and/or to try out a new release tag without
creating a proper version .oe file.

Note that $PV is still picked up from the .oe file, and through that the
sysdeps files we use, so this cannot be used if the sysdeps format has
changed too much.