openresolv: new recipe
We amend the included resolvconf.conf by disabling all hooks except
libc, since most embedded systems probably don't have the software in
question anyway, and for the libc hook we avoid doing a lot of work to
figure out how to restart a probably-not-even-existing nscd service.

The user of this recipe may of course provide his own openresolv-config
overriding all of that.