openssh: allow default ssh session PATH to be set via USE flags
By default, openssh set a PATH of /usr/bin:/bin:/usr/sbin:/sbin for a
new session. This is different from the default busybox value of
/sbin:/usr/sbin:/bin:/usr/bin .

The value used by openssh can be set at configure time with the
--with-default-path option, and if a different value is required for
uid 0, --with-superuser-path. We can easily allow e.g. the distro to
set these values by providing USE flags for them, the absence of which
will just keep the current behaviour.