class/s6rc-sanity: Fix pipeline dependencies on dynamic services
In order to allow statically defined services (i.e. in /etc/rc) to have a
producer-for/consumer-for relation to a dynamic service (i.e. one created by a
hook in /etc/rc.hooks/), we create dummy longrun services instead of empty
bundles, and take care to add needed producer-for/consumer-for relations.

Hopefully the hooks will also do the right thing.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>