s6-init: Allow blocking mode to service command
Adding a '-b' option to service command allows blocking on the s6-rc lock
starting and stopping services.  Without this, service command fails when
called while another (s6-rc) program holds the lock, spewing error messages
like

    s6-rc: fatal: unable to take locks: Resource temporarily unavailable

This is generally the safe approach, so it is kept as the default.

But when using service command while changing s6-rc state (for example during
boot), it can be convenient to wait for the s6-rc lock to be released.

A valid use-case is when an application (that is not running as an s6-rc
service!) wants to stop an s6-rc service.  To avoid failure when started
during boot, the -b option is very helpful.

Note: The -b option must not be used in nested s6-rc invocation.  Anything
running with the the s6-rc lock held (for example an s6-rc service script)
will deadlock forever if using service command with -b option!

Signed-off-by: Esben Haabendal <esben@haabendal.dk>